Build rdev Site

build_rdev_site() is a wrapper for pkgdown::build_site_github_pages() optimized for rdev workflow that updates README.md and performs a clean GitHub pages build using pkgdown.

Usage

build_rdev_site(pkg = ".", ...)

Arguments

pkg

Path to package. Currently, only pkg = "." is supported.

...

additional arguments passed to pkgdown::build_site_github_pages() (not implemented)

Details

When run, build_rdev_site() calls:

devtools::build_readme()

pkgdown::build_site_github_pages() with install = TRUE and new_process = TRUE

Continuous Integration

Both build_rdev_site() and build_analysis_site() are meant to be used as part of a CI/CD workflow, and temporarily set the environment variable CI == "TRUE" so that the build will fail when non-internal topics are not included on the reference index page per pkgdown::build_reference().