Initialize rdev package

Initialize a rdev package within a newly created create_github_repo() project by creating a new git branch, committing all changes, and running use_rdev_package().

Usage

init()

Details

init() will stop if rlang::is_interactive() is FALSE.

After running init(), update the Title and Description fields in the DESCRIPTION file without committing and run either setup_analysis() or setup_rdev() per the quickstart.

See also