Lint all files

Lint all files in a project. Implemented as a wrapper for lintr::lint_dir().

lint_all(path = ".", ...)

Arguments

path

the path to the base directory, by default, it will be searched in the parent directories of the current directory.

...

Arguments passed on to lintr::lint_dir

relative_path

if TRUE, file paths are printed using their path relative to the base directory. If FALSE, use the full absolute path.

exclusions

exclusions for exclude, relative to the package path.

pattern

pattern for files, by default it will take files with any of the extensions .R, .Rmd, .Rnw, .Rhtml, .Rrst, .Rtex, .Rtxt allowing for lowercase r (.r, ...)

parse_settings

whether to try and parse the settings.

Value

A list of lint objects.

Examples

if (FALSE) {
lint_all()
lint_all("analysis")
}