Convert R Notebook to html_document

Copies a file using fs::file_copy(), and changes the output type in the yaml front matter from html_notebook to html_document. If the file is not type html_notebook, it is copied without changing the output type.

to_document(file_path, new_path, overwrite = FALSE)

Arguments

file_path
A string path to the source file

new_path
A string path to copy the converted file using fs::file_copy()

overwrite
Overwrite files if they exist, passed to fs::file_copy()

Details

Warning: to_document() is currently considered Experimental.

Examples

if (FALSE) {
to_document("notebook.Rmd", "document.Rmd")
to_document("notebooks_dir/notebook.Rmd", "documents_dir")
}