Update WORDLIST from notebooks

Update package inst/WORDLIST with words from spell_check_notebooks().

Usage

update_wordlist_notebooks(
  pkg = ".",
  vignettes = TRUE,
  path = "analysis",
  regexp = "[.][Rq]md$",
  confirm = TRUE
)

Arguments

pkg

path to package root directory containing the DESCRIPTION file

vignettes

check all rmd and rnw files in the pkg root directory (e.g. readme.md) and package vignettes folder.

path

A character vector of one or more paths.

regexp

A regular expression (e.g. [.]csv$) passed on to grep() to filter paths.

confirm

show changes and ask confirmation before adding new words to the list

Details

update_wordlist_notebooks is a customized version of spelling::update_wordlist() using code from wordlist.R.