Build Analysis Site

build_analysis_site() is a wrapper for pkgdown::build_site() that adds an 'Analysis' menu containing rendered versions of all .Rmd files in analysis/.

build_analysis_site(pkg = ".", ...)

Arguments

pkg
Path to package.

...
additional arguments passed to pkgdown::build_site()

Details

When run, build_analysis_site():

Reads base pkgdown settings from pkgdown/_base.yml

Writes base settings to pkgdown/_pkgdown.yml

Creates a template using pkgdown::template_navbar() and inserts an analysis menu with links to html versions of each .Rmd file in analysis/

Writes the template to pkgdown/_pkgdown.yml

Runs pkgdown::clean_site() and pkgdown::build_site()

Creates a _site.yml file based on the final _pkgdown.yml that clones the pkgdown navbar in a temporary build directory

Copies the following from analysis/ into the build directory: *.Rmd, assets/, data/, import/, rendered/

Changes *.Rmd from html_notebook to html_document using to_document()

Builds a site using rmarkdown::render_site() using modified html_document output settings to render files with the look and feel of html_notebook

Moves the rendered files to docs/: *.html, assets/, rendered/, without overwriting

build_analysis_site() will fail with an error if there are no files in analysis/*.Rmd, or if pkgdown/_base.yml does not exist.

Warning: build_analysis_site() is currently considered Experimental. Currently only build_analysis_site(pkg = ".") is supported.

Examples

if (FALSE) {
build_analysis_site()
}