Build rdev Site

build_rdev_site() is a wrapper for pkgdown::build_site() optimized for rdev workflow that updates README.md and performs a clean build using pkgdown.

build_rdev_site(pkg = ".", ...)

Arguments

pkg

Path to package. Currently, only pkg = "." is supported.

...

additional arguments passed to pkgdown::build_site() (not implemented)

Details

When run, build_rdev_site() calls:

devtools::build_readme()

pkgdown::clean_site()

pkgdown::build_site()

Continuous Integration

Both build_rdev_site() and build_analysis_site() are meant to be used as part of a CI/CD workflow, and temporarily set the environment variable CI == "TRUE" so that the build will fail when non-internal topics are not included on the reference index page per pkgdown::build_reference().