Stage a GitHub release

Open a GitHub pull request for a new release from NEWS.md. Approve, merge, and create the release using merge_release().

stage_release(pkg = ".", filename = "NEWS.md", host = NULL)

Arguments

pkg

path to package. Currently, only pkg = "." is supported.

filename

name of file containing release notes, defaults to NEWS.md.

host

GitHub host to target, passed to the .api_url argument of gh::gh(). If unspecified, gh defaults to "https://api.github.com", although gh's default can be customised by setting the GITHUB_API_URL environment variable.

For a hypothetical GitHub Enterprise instance, either "https://github.acme.com/api/v3" or "https://github.acme.com" is acceptable.

Value

results of GitHub pull request, invisibly

Details

When run, stage_release():

Extracts release version and release notes from NEWS.md using get_release()

Validates version conforms to rdev conventions (#.#.#) and release notes aren't empty

Verifies that version tag doesn't already exist using gert::git_tag_list()

Checks for uncommitted changes and stops if any exist using gert::git_diff_patch()

Creates a new branch if on the default branch (gert::git_branch() == usethis::git_default_branch()) using gert::git_branch_create()

Updates Version in DESCRIPTION with desc::desc_set_version(), commits and push to git with message "GitHub release <version>" using gert::git_add(), gert::git_commit() and gert::git_push()

Runs build_analysis_site() (if pkgdown/_base.yml exists) or build_rdev_site(), commits and pushes changes to git with message "<builder> for release <version>"

Opens a pull request with the title "<package> <version>" and the release notes in the body using gh::gh()