Functions re-exported from the urlchecker package

These functions are re-exported from the urlchecker package.

Usage

url_check(
  path = ".",
  db = NULL,
  parallel = TRUE,
  pool = curl::new_pool(),
  progress = TRUE
)

url_update(path = ".", results = url_check(path))

Arguments

path

Path to the package

db

A url database

parallel

If TRUE, check the URLs in parallel

pool

A multi handle created by curl::new_pool(). If NULL use a global pool.

progress

Whether to show the progress bar for parallel checks

results

results from url_check.

Details

Follow the links below to see the urlchecker documentation.

urlchecker::url_check()

urlchecker::url_update()